+++ src/main/java/com/connor/hozon/bom/bomSystem/dao/fullCfg/HzFullCfgWithCfgDao.java
@@ -11,7 +11,6 @@ import com.connor.hozon.bom.bomSystem.dao.BasicDao;
 import org.springframework.context.annotation.Configuration;
 import sql.pojo.cfg.fullCfg.HzFullCfgWithCfg;
 
-import java.math.BigDecimal;
 import java.util.List;
 /**
  * @Author: Fancyears·Maylos·Maywas
@@ -28,7 +27,7 @@ public interface HzFullCfgWithCfgDao extends BasicDao<HzFullCfgWithCfg> {
 
     HzFullCfgWithCfg selectByPrimaryKey(Long id);
 
-    List<HzFullCfgWithCfg> selectByMainID(BigDecimal flCfgVersion);
+    List<HzFullCfgWithCfg> selectByMainID(Long flCfgVersion);
 
     int insertBomLine(List<HzFullCfgWithCfg> hzFullCfgWithCfgs);
 
@@ -36,7 +35,7 @@ public interface HzFullCfgWithCfgDao extends BasicDao<HzFullCfgWithCfg> {
 
     List<HzFullCfgWithCfg> query2YCfgByProjectId(String projectId);
 
-    HzFullCfgWithCfg selectByBomLineUidWithVersion(BigDecimal version, String puid);
+    HzFullCfgWithCfg selectByBomLineUidWithVersion(Long version, String puid);
 
     HzFullCfgWithCfg query2YCfgByBomLineId(String bomLineId);
 
@@ -47,7 +46,7 @@ public interface HzFullCfgWithCfgDao extends BasicDao<HzFullCfgWithCfg> {
      * @param puid
      * @return
      */
-    HzExFullCfgWithCfg selectByBLOutWithCfg(BigDecimal version, String puid);
+    HzExFullCfgWithCfg selectByBLOutWithCfg(Long version, String puid);
     /**
      * 根据2Y和全配主键一起查询，连带出特性值对象和BOMLine对象
      *
@@ -55,6 +54,6 @@ public interface HzFullCfgWithCfgDao extends BasicDao<HzFullCfgWithCfg> {
      * @param puid
      * @return
      */
-    HzExFullCfgWithCfg selectByBLOutWithCfgAndBL(BigDecimal version, String puid);
+    HzExFullCfgWithCfg selectByBLOutWithCfgAndBL(Long version, String puid);
 
 }
